### 2021-08-24 #75 @whedon
[PRE REVIEW]: DeconvOptim.jl - Signal Deconvolution with Julia
**Submitting author:** @roflmaostc (<a href="http://orcid.org/0000-0001-8763-2948">Felix Wechsler</a>)
**Repository:** <a href="https://github.com/roflmaostc/DeconvOptim.jl" target ="_blank">https://github.com/roflmaostc/DeconvOptim.jl</a>
**Version:**
