### 2021-08-24 #75 @roflmaostc
Thanks for activating the review!

It looks like that PDF building has failed because of the missing `pgfplots.sty`. Usually, that's included with the normal LaTeX distributions, so I'm not quite sure what's installed on the backend and why that's missing.
