### 2021-09-03 #75 @roflmaostc
Anything I can help with to fix the LaTeX errors?
However, similar errors are also in  [the other review](https://github.com/JuliaCon/proceedings-review/issues/76).
