### 2021-09-15 #75 @matbesancon
@roflmaostc it seems it's not included. The easiest solution for now would be to add the sty file in your project paper folder directly
