### 2021-09-15 6d3ead5
Update latex header [no ci]
