+++ paper/header.tex
@@ -1,18 +1,19 @@
 % **************GENERATED FILE, DO NOT EDIT**************
 
-\title{My JuliaCon proceeding}
+\title{DeconvOptim.jl - Signal Deconvolution with Julia}
 
-\author[1]{1st author}
-\author[1, 2]{2nd author}
-\author[2]{3rd author}
-\affil[1]{University}
-\affil[2]{National Lab}
+\author[1, 2]{Felix Wechsler}
+\author[1, 2, 3]{Rainer Heintzmann}
+\affil[1]{Faculty of Physics and Astronomy, Friedrich-Schiller-University, Jena, Germany}
+\affil[2]{Leibniz Institute of Photonic Technology, Albert-Einstein-Straße 9, 07745 Jena, Germany}
+\affil[3]{Institute of Physical Chemistry and Abbe Center of Photonics, Friedrich-Schiller-University, Helmholtzweg 4, Jena, Germany}
 
-\keywords{Julia, Optimization, Game theory, Compiler}
+\keywords{Julia, Image Processing, Deconvolution, Microscopy}
 
 \hypersetup{
-pdftitle = {My JuliaCon proceeding},
+pdftitle = {DeconvOptim.jl - Signal Deconvolution with Julia},
 pdfsubject = {JuliaCon 2019 Proceedings},
-pdfauthor = {1st author, 2nd author, 3rd author},
-pdfkeywords = {Julia, Optimization, Game theory, Compiler},
+pdfauthor = {Felix Wechsler, Rainer Heintzmann},
+pdfkeywords = {Julia, Image Processing, Deconvolution, Microscopy},
 }
+
