+++ paper/siunitx-abbreviations.cfg
@@ -4,7 +4,7 @@
 %%
 %% The original source files were:
 %%
-%% siunitx-v2.dtx  (with options: `config,abbreviations')
+%% siunitx.dtx  (with options: `config,abbreviations')
 %% ---------------------------------------------------------------
 %% The siunitx package --- A comprehensive (SI) units package
 %% Maintained by Joseph Wright
@@ -13,7 +13,7 @@
 %% See http://www.latex-project.org/lppl.txt
 %% ---------------------------------------------------------------
 %% 
-\ProvidesExplFile {siunitx-abbreviations.cfg} {2021-04-17} {2.8e}
+\ProvidesExplFile {siunitx-abbreviations.cfg} {2017/11/26} {2.7k}
   {siunitx: Abbreviated units}
 \DeclareSIUnit \A  {        \ampere }
 \DeclareSIUnit \pA { \pico  \ampere }
@@ -104,7 +104,7 @@
 \DeclareSIUnit \ns { \nano  \second }
 \DeclareSIUnit \us { \micro \second }
 \DeclareSIUnit \ms { \milli \second }
-%% Copyright (C) 2008-2021 by
+%% Copyright (C) 2008-2019 by
 %%   Joseph Wright <joseph.wright@morningstar2.co.uk>
 %% 
 %% It may be distributed and/or modified under the conditions of
