+++ paper/siunitx-version-1.cfg
@@ -4,7 +4,7 @@
 %%
 %% The original source files were:
 %%
-%% siunitx-v2.dtx  (with options: `config,version-1')
+%% siunitx.dtx  (with options: `config,version-1')
 %% ---------------------------------------------------------------
 %% The siunitx package --- A comprehensive (SI) units package
 %% Maintained by Joseph Wright
@@ -13,7 +13,7 @@
 %% See http://www.latex-project.org/lppl.txt
 %% ---------------------------------------------------------------
 %% 
-\ProvidesExplFile {siunitx-version-1.cfg} {2021-04-17} {2.8e}
+\ProvidesExplFile {siunitx-version-1.cfg} {2017/11/26} {2.7k}
   {siunitx: Version 1 settings}
 \keys_define:nn { siunitx } {
   addsign        .choice: ,
@@ -381,21 +381,18 @@
       {
         multi-part-units      = single,
         product-units         = single,
-        range-units           = single
       },
   repeatunits /
     power        .meta:n    =
       {
         multi-part-units      = repeat,
         product-units         = power,
-        range-units           = repeat
       },
   repeatunits /
     true        .meta:n       =
       {
         multi-part-units      = repeat,
         product-units         = repeat,
-        range-units           = repeat
       },
   repeatunits    .default:n = { true                                  },
   retainplus     .meta:n    = { retain-explicit-plus    = #1          },
@@ -730,7 +727,7 @@
 \DeclareSIUnit \yb        { \yocto \barn }
 \NewDocumentCommand \requiresiconfigs { m }
   { \keys_set:nn { siunitx } { version-1-compatibility } }
-%% Copyright (C) 2008-2021 by
+%% Copyright (C) 2008-2019 by
 %%   Joseph Wright <joseph.wright@morningstar2.co.uk>
 %% 
 %% It may be distributed and/or modified under the conditions of
@@ -748,3 +745,4 @@
 %%                                 siunitx.ins.
 %%
 %% End of file `siunitx-version-1.cfg'.
+
