+++ paper/siunitx-binary.cfg
@@ -4,7 +4,7 @@
 %%
 %% The original source files were:
 %%
-%% siunitx-v2.dtx  (with options: `config,binary')
+%% siunitx.dtx  (with options: `config,binary')
 %% ---------------------------------------------------------------
 %% The siunitx package --- A comprehensive (SI) units package
 %% Maintained by Joseph Wright
@@ -13,7 +13,7 @@
 %% See http://www.latex-project.org/lppl.txt
 %% ---------------------------------------------------------------
 %% 
-\ProvidesExplFile {siunitx-binary.cfg} {2021-04-17} {2.8e}
+\ProvidesExplFile {siunitx-binary.cfg} {2017/11/26} {2.7k}
   {siunitx: Binary units}
 \DeclareBinaryPrefix \kibi { Ki } { 10 }
 \DeclareBinaryPrefix \mebi { Mi } { 20 }
@@ -25,7 +25,7 @@
 \DeclareBinaryPrefix \yobi { Yi } { 80 }
 \DeclareSIUnit \bit  { bit }
 \DeclareSIUnit \byte { B }
-%% Copyright (C) 2008-2021 by
+%% Copyright (C) 2008-2019 by
 %%   Joseph Wright <joseph.wright@morningstar2.co.uk>
 %% 
 %% It may be distributed and/or modified under the conditions of
@@ -43,3 +43,4 @@
 %%                                 siunitx.ins.
 %%
 %% End of file `siunitx-binary.cfg'.
+
