### 2021-09-22 f0f6dc4
Remove additional file so that paper.tex compiles again
