+++ paper/paper.tex
@@ -9,6 +9,7 @@
 \usepackage{amsmath}
 \usepackage{tikz}
 \usepackage{pgfplots}
+% \pgfplotsset{compat=1.17}
 \usepackage{siunitx}
 \usepackage{here}
 \newcommand\todo[1]{\textcolor{red}{\textbf{TODO: #1}}}
