+++ paper/siunitx-binary.cfg
@@ -1,46 +1,0 @@
-%%
-%% This is file `siunitx-binary.cfg',
-%% generated with the docstrip utility.
-%%
-%% The original source files were:
-%%
-%% siunitx.dtx  (with options: `config,binary')
-%% ---------------------------------------------------------------
-%% The siunitx package --- A comprehensive (SI) units package
-%% Maintained by Joseph Wright
-%% E-mail: joseph.wright@morningstar2.co.uk
-%% Released under the LaTeX Project Public License v1.3c or later
-%% See http://www.latex-project.org/lppl.txt
-%% ---------------------------------------------------------------
-%% 
-\ProvidesExplFile {siunitx-binary.cfg} {2017/11/26} {2.7k}
-  {siunitx: Binary units}
-\DeclareBinaryPrefix \kibi { Ki } { 10 }
-\DeclareBinaryPrefix \mebi { Mi } { 20 }
-\DeclareBinaryPrefix \gibi { Gi } { 30 }
-\DeclareBinaryPrefix \tebi { Ti } { 40 }
-\DeclareBinaryPrefix \pebi { Pi } { 50 }
-\DeclareBinaryPrefix \exbi { Ei } { 60 }
-\DeclareBinaryPrefix \zebi { Zi } { 70 }
-\DeclareBinaryPrefix \yobi { Yi } { 80 }
-\DeclareSIUnit \bit  { bit }
-\DeclareSIUnit \byte { B }
-%% Copyright (C) 2008-2019 by
-%%   Joseph Wright <joseph.wright@morningstar2.co.uk>
-%% 
-%% It may be distributed and/or modified under the conditions of
-%% the LaTeX Project Public License (LPPL), either version 1.3c of
-%% this license or (at your option) any later version.  The latest
-%% version of this license is in the file:
-%%    http://www.latex-project.org/lppl.txt
-%% 
-%% This work is "maintained" (as per LPPL maintenance status) by
-%%   Joseph Wright.
-%% 
-%% This work consists of the file  siunitx.dtx
-%%           and the derived files siunitx.pdf,
-%%                                 siunitx.sty and
-%%                                 siunitx.ins.
-%%
-%% End of file `siunitx-binary.cfg'.
-
