+++ paper/siunitx-version-1.cfg
@@ -1,748 +1,0 @@
-%%
-%% This is file `siunitx-version-1.cfg',
-%% generated with the docstrip utility.
-%%
-%% The original source files were:
-%%
-%% siunitx.dtx  (with options: `config,version-1')
-%% ---------------------------------------------------------------
-%% The siunitx package --- A comprehensive (SI) units package
-%% Maintained by Joseph Wright
-%% E-mail: joseph.wright@morningstar2.co.uk
-%% Released under the LaTeX Project Public License v1.3c or later
-%% See http://www.latex-project.org/lppl.txt
-%% ---------------------------------------------------------------
-%% 
-\ProvidesExplFile {siunitx-version-1.cfg} {2017/11/26} {2.7k}
-  {siunitx: Version 1 settings}
-\keys_define:nn { siunitx } {
-  addsign        .choice: ,
-  addsign /
-    all          .meta:n    = { explicit-sign           = +           },
-  addsign /
-    exp          .meta:n    = { explicit-sign           = +           },
-  addsign /
-    false        .meta:n    = { explicit-sign           =             },
-  addsign /
-    mant         .meta:n    = { explicit-sign           = +           },
-  addsign /
-    none         .meta:n    = { explicit-sign           =             },
-  addsign /
-    true         .meta:n    = { explicit-sign           = +           },
-  addsign        .default:n = { true                                  },
-  allowlitunits  .meta:n    = { free-standing-units     = #1          },
-  allowlitunits  .default:n = { true                                  },
-  allowoptarg    .meta:n    = { unit-optional-argument  = #1          },
-  allowoptargs   .default:n = { true                                  },
-  allowzeroexp   .meta:n    = { retain-zero-exponent    = #1          },
-  allowzeroexp   .default:n = { true                                  },
-  alsoload       .code:n    = ,
-  angformat      .code:n    = ,
-  anglesep       .choice: ,
-  anglesep /
-    cdot         .meta:n    =
-      { arc-separator         = \ensuremath { { } \cdot { } } },
-  anglesep /
-    comma        .meta:n    = { arc-separator           = \ensuremath { { {,} } } },
-  anglesep /
-    fullstop     .meta:n    = { arc-separator           = \ensuremath { { { . } } } },
-  anglesep /
-    med          .meta:n    = { arc-separator           = \ensuremath { \: } },
-  anglesep /
-    medium       .meta:n    = { arc-separator           = \ensuremath { \: } },
-  anglesep /
-    none         .meta:n    = { arc-separator           =             },
-  anglesep /
-    period       .meta:n    = { arc-separator           = \ensuremath { { { . } } } },
-  anglesep /
-    space        .meta:n    = { arc-separator           = \text { ~ } },
-  anglesep /
-    stop         .meta:n    = { arc-separator           = \ensuremath { { { . } } } },
-  anglesep /
-    thick        .meta:n    = { arc-separator           = \ensuremath { \; } },
-  anglesep /
-    thin         .meta:n    = { arc-separator           = \ensuremath { \, }  },
-  anglesep /
-    tightcdot    .meta:n    =
-      { arc-separator           = \ensuremath { \bgroup \cdot \egroup } },
-  anglesep /
-    tighttimes   .meta:n    =
-      { arc-separator           = \ensuremath { \bgroup \times \egroup } },
-  anglesep /
-    times        .meta:n    = { arc-separator           = \ensuremath { \times } },
-  anglesep /
-    unknown      .meta:n    = { arc-separator           = \ensuremath {#1} },
-  astroang       .meta:n    =
-    { angle-symbol-over-decimal = #1                                  },
-  astroang       .default:n = { true                                  },
[…]
