+++ paper/siunitx-abbreviations.cfg
@@ -1,124 +1,0 @@
-%%
-%% This is file `siunitx-abbreviations.cfg',
-%% generated with the docstrip utility.
-%%
-%% The original source files were:
-%%
-%% siunitx.dtx  (with options: `config,abbreviations')
-%% ---------------------------------------------------------------
-%% The siunitx package --- A comprehensive (SI) units package
-%% Maintained by Joseph Wright
-%% E-mail: joseph.wright@morningstar2.co.uk
-%% Released under the LaTeX Project Public License v1.3c or later
-%% See http://www.latex-project.org/lppl.txt
-%% ---------------------------------------------------------------
-%% 
-\ProvidesExplFile {siunitx-abbreviations.cfg} {2017/11/26} {2.7k}
-  {siunitx: Abbreviated units}
-\DeclareSIUnit \A  {        \ampere }
-\DeclareSIUnit \pA { \pico  \ampere }
-\DeclareSIUnit \nA { \nano  \ampere }
-\DeclareSIUnit \uA { \micro \ampere }
-\DeclareSIUnit \mA { \milli \ampere }
-\DeclareSIUnit \kA { \kilo  \ampere }
-\DeclareSIUnit \Hz  {        \hertz }
-\DeclareSIUnit \mHz { \milli \hertz }
-\DeclareSIUnit \kHz { \kilo  \hertz }
-\DeclareSIUnit \MHz { \mega  \hertz }
-\DeclareSIUnit \GHz { \giga  \hertz }
-\DeclareSIUnit \THz { \tera  \hertz }
-\DeclareSIUnit \mol  {        \mole }
-\DeclareSIUnit \fmol { \femto \mole }
-\DeclareSIUnit \pmol { \pico  \mole }
-\DeclareSIUnit \nmol { \nano  \mole }
-\DeclareSIUnit \umol { \micro \mole }
-\DeclareSIUnit \mmol { \milli \mole }
-\DeclareSIUnit \kmol { \kilo  \mole }
-\DeclareSIUnit \V  {        \volt }
-\DeclareSIUnit \pV { \pico  \volt }
-\DeclareSIUnit \nV { \nano  \volt }
-\DeclareSIUnit \uV { \micro \volt }
-\DeclareSIUnit \mV { \milli \volt }
-\DeclareSIUnit \kV { \kilo  \volt }
-\DeclareSIUnit \hl { \hecto \litre }
-\DeclareSIUnit \l  {        \litre }
-\DeclareSIUnit \ml { \milli \litre }
-\DeclareSIUnit \ul { \micro \litre }
-\DeclareSIUnit \hL { \hecto \liter }
-\DeclareSIUnit \L  {        \liter }
-\DeclareSIUnit \mL { \milli \liter }
-\DeclareSIUnit \uL { \micro \liter }
-\DeclareSIUnit \fg  { \femto \gram }
-\DeclareSIUnit \pg  { \pico  \gram }
-\DeclareSIUnit \ng  { \nano  \gram }
-\DeclareSIUnit \ug  { \micro \gram }
-\DeclareSIUnit \mg  { \milli \gram }
-\DeclareSIUnit \g   { \gram }
-\DeclareSIUnit \kg  { \kilo \gram }
-\DeclareSIUnit \amu { \atomicmassunit }
-\DeclareSIUnit \W   {        \watt }
-\DeclareSIUnit \uW  { \micro \watt }
-\DeclareSIUnit \mW  { \milli \watt }
-\DeclareSIUnit \kW  { \kilo  \watt }
-\DeclareSIUnit \MW  { \mega  \watt }
-\DeclareSIUnit \GW  { \giga  \watt }
-\DeclareSIUnit \J   { \joule }
-\DeclareSIUnit \uJ  { \micro \joule }
[…]
