+++ Project.toml
@@ -1,7 +1,7 @@
 name = "DeconvOptim"
 uuid = "03e7cd2f-1a03-4ea9-b59b-760a446df67f"
 authors = ["Felix Wechsler <fxw+github@mailbox.org>"]
-version = "0.5.0"
+version = "0.5.1"
 
 [deps]
 ChainRulesCore = "d360d2e6-b24c-11e9-a2a3-2a2ae2dbcce4"
@@ -20,8 +20,8 @@ Zygote = "e88e6eb3-aa80-5325-afca-941959d7151f"
 
 [compat]
 ChainRulesCore = "1"
-FillArrays = "0.12" 
 FFTW = "1.2, 1.3, 1.4"
+FillArrays = "0.12"
 Interpolations = "0.12.10, 0.13"
 LineSearches = "7.1"
 Noise = "0.3.0"
