### 2021-09-22 2bf19e9
Fix rrules for loss functions
