+++ Project.toml
@@ -25,11 +25,11 @@ FillArrays = "0.12"
 Interpolations = "0.12.10, 0.13"
 LineSearches = "7.1"
 Noise = "0.3.0"
-Optim = "1.2, 1.3, 1.4"
+Optim = "1.2, 1.3, 1.4, 1.5"
 Requires = "1.1.0"
 StatsBase = "0.33"
 Tullio = "0.3"
-Zygote = "0.6.19"
+Zygote = "0.6"
 julia = "1.5, 1.6"
 
 [extras]
