### 2022-01-05 551462b
added the option to keep the large size by using a negative padding.
