### 2022-01-09 bf639ae
added `options_trace_deconv` in analysis tools, which eases comparative evaluation of deconvolutions. Also added an example.
