### 2022-01-09 db3e968
updated test and introduced the `initial` argument to `deconvolution`
