### 2022-01-12 b07980f
updates and fixes. Changed the return type of options_trace_deconv
