+++ src/analysis_tools.jl
@@ -98,16 +98,28 @@ Calculates the mean variance between two array, but normalizing arra a to the sa
 function normalized_variance(a, b)
     factor = sum(b)/sum(a)
     sum(abs2.(a.*factor .-b))./prod(size(a))
-    # sum(abs2.(a./sum(a).*sum(b) .-b))
 end
 
+function reset_summary!(summary)
+    summary["losses"] = []
+    summary["best_ncc"] = -Inf
+    summary["best_ncc_idx"] = 0
+    summary["best_ncc_img"] = []
+    summary["nccs"] = []
+    summary["best_nvar"] = Inf
+    summary["best_nvar_idx"] = 0
+    summary["best_nvar_img"] = []
+    summary["times"] = []
+    summary["step_sizes"] = []
+    summary["nvars"] = []
+end
 
 """
     options_trace_deconv(ground_truth, iterations, mapping, every=1)
 
     A useful routine to simplify performance checks of deconvolution on simulated data.
     Returns an Options structure to be used with the deconvolution routine as an argument to `opt_options` and 
-    a function that returns the summary dictionary with all the performance metrics calculated.
+    a summary dictionary with all the performance metrics calculated, which is resetted and updated during deconvolution.
     This can then be plotted or visualized.
     The summary dictionary has the following content:
 
@@ -123,10 +135,13 @@ end
 
     For an example of how to plot the results, see the file `` in the `examples` folder.
 # Arguments
-- `ground_truth`: The underlying ground truth data. Note that this scaling is unimportant due to the normalized norms used for comparison, whereas the relative offset matters. 
+- `ground_truth`: The underlying ground truth data. Note that this scaling is unimportant due to the normalized norms used for comparison, 
+                  whereas the relative offset matters. 
 - `iterations`: The maximal number of iterations to performance. If covergence is reached, the result may have less iterations
-- `mapping`: If mappings such as the positivity contraints (e.g. `NonNegative()`) are used in the deconvolution routing, they also need to be provided here. Otherwises select `nothing`.
-- `every`: This option allows to select every how many iterations the evaluation is performed. Note that the results will not keep track of this iteration number. 
+- `mapping`: If mappings such as the positivity contraints (e.g. `NonNegative()`) are used in the deconvolution routing, they also 
+             need to be provided here. Otherwises select `nothing`.
+- `every`: This option allows to select every how many iterations the evaluation is performed. Note that the results will not keep track 
+        of this iteration number. 
 
 # Example
 ```julia-repl
@@ -168,56 +183,47 @@ julia> @vt (ground_truth, best_ncc_img, best_nvar_img) = show_noreg(true)
 ```
 """
 function options_trace_deconv(ground_truth, iterations, mapping, every=1)
-    summary = Dict(
-        "losses" => [],
-        "best_ncc" => -Inf,
-        "best_ncc_idx" => 0,
-        "best_ncc_img" => [],
-        "nccs" => [],
-        "best_nvar" => Inf,
-        "best_nvar_idx" => 0,
-        "best_nvar_img" => [],
-        "nvars" => [] )
+    summary = Dict()
+    reset_summary!(summary)
     idx = 1
     cb = tr -> begin
-            if tr[end].iteration == 1
-                    summary["losses"] = [];summary["best_ncc"] = -Inf;summary["best_ncc_idx"] = 0;summary["best_ncc_img"] = [];summary["nccs"] = []
-                    summary["best_nvar"] = Inf;summary["best_nvar_idx"] = 0;summary["best_nvar_img"] = [];summary["nvars"] = []
-                    idx = 1
-            end
-            loss = tr[end].value
-            push!(summary["losses"], loss)
[…]
