### 2022-01-12 d9520f3
fixed the error check for providing the iteration argument
