### 2022-01-12 c7ffbd5
changed the default for Anscombe to 0 as it works better and makes more sense for unnormalized data.
