### 2022-01-15 7710555
added varargs for additional arguments to be passed to the optim structure
