### 2022-01-16 21f212f
modified the callback mechanism to also work with the Richardson-Lucy iterations
