+++ Project.toml
@@ -1,7 +1,7 @@
 name = "DeconvOptim"
 uuid = "03e7cd2f-1a03-4ea9-b59b-760a446df67f"
 authors = ["Felix Wechsler <fxw+github@mailbox.org>"]
-version = "0.5.1"
+version = "0.6.0"
 
 [deps]
 ChainRulesCore = "d360d2e6-b24c-11e9-a2a3-2a2ae2dbcce4"
@@ -24,13 +24,13 @@ FFTW = "1.2, 1.3, 1.4"
 FillArrays = "0.12"
 Interpolations = "0.12.10, 0.13"
 LineSearches = "7.1"
-Noise = "0.3.0"
-Optim = "1.2, 1.3, 1.4, 1.5"
+Noise = "0.3"
+Optim = "1.2, 1.3, 1.4, 1.5, 1.6"
 Requires = "1.1.0"
 StatsBase = "0.33"
 Tullio = "0.3"
 Zygote = "0.6"
-julia = "1.5, 1.6"
+julia = "1.5, 1.6, 1.7"
 
 [extras]
 ImageMagick = "6218d12a-5da1-5696-b52f-db25d2ecc6d1"
