+++ Project.toml
@@ -21,7 +21,7 @@ Zygote = "e88e6eb3-aa80-5325-afca-941959d7151f"
 [compat]
 ChainRulesCore = "1"
 FFTW = "1.2, 1.3, 1.4"
-FillArrays = "0.12"
+FillArrays = "0.12, 0.13"
 Interpolations = "0.12.10, 0.13"
 LineSearches = "7.1"
 Noise = "0.3"
