### 2022-05-06 10c75d3
Change default deconvolution code to no regularizer
