### 2022-06-30 #99 @whedon
[REVIEW]: DeconvOptim.jl - Signal Deconvolution with Julia
**Submitting author:** @roflmaostc (<a href="http://orcid.org/0000-0001-8763-2948">Felix Wechsler</a>)
**Repository:** <a href="https://github.com/roflmaostc/DeconvOptim.jl" target ="_blank">https://github.com/roflmaostc/DeconvOptim.jl</a>
**Version:** v0.7.1
