### 2022-08-11 #99 @sloede
### Summary of the paper
The manuscript describes the software package *DeconvOptim.jl*, a toolbox written in Julia for the deconvolution of measured signals that have been degraded by blurring and/or Gaussian noise/Poisson shot noise. It gives an introduction to the pertinent basics of digital signal processing based on deconvolution, and discusses some of the common design choices and issues when using this approach. It then describes the pipeline of the deconvolution approach used in DeconvOptim.jl. Finally, experiments with one- and multi-dimensional data sets are shown, where an emphasis is put on the influence of regularization. Further, performance comparisons to existing deconvolution tools are shown, demonstrating that the presented software package is ahead of or on par with most of the other tools.

### Review
Overall, the paper is well written, with a clear storyline, a helpful introduction for readers not familiar with deconvolution techniques, and useful examples in the experiment section. There are only a few minor issues that I noted. Once these are adequately addressed, I think this paper can be accepted for publication (*minor modification request*):

a) p. 1, sec. 2: "...we also need **an** efficient way..."

b) p. 3, bottom right: It is unfortunate that the code listing is split across two pages. I strongly recommend to reformat its placement such that it is shown completely on a single page.

c) p. 4, sec. 5: The section introduction once mentions Julia 1.6.1 and once Julia 1.6.2, thus it is not clear which version was used.

d) p. 4, sec. 5.2: "Figure 5c shows the result after 50 iterations..." --> I believe this should read "Figure **5b** shows the..."

e) p. 4, sec. 5.2: "We see [...] heavily affected by some artifacts [...]" and later "...it can be seen visually as well.": In both cases, it is not clear to me what exactly these artifacts are and what "it" is that can be seen visually. Since this is a qualitative analysis and not a quantitative one, I think the authors should offer considerably more detail on which particular features they make their observations that one approach is better than the other. Further, the qualitative analysis should be augmented by showing the original ("ground truth") image in Fig. 5 along with the convolved images.

f) p. 4, fig. 4 *and* p. 5, fig. 6: Both these images seem to have issues with the correct labeling of the axes and the legends (i.e., symbols and digits are missing). As they are right now, it is hard to fully understand them.

g) p. 5, sec. 5.3: It is not obvious to me why it is a four-dimensional dataset when talking about a 3D image. If it the color channels represent the fourth dimension, I suggest to explicitly say so.

h) p. 5, sec. 5.4.1: "...being able to use ~~of~~ all threads...": I think the "of" is erroneous and should be removed.

i) p. 6, tab. 1: It is not clear from the table description alone what the difference is between the regular measurements and the ones with the nabla sign prefixed. This should be improved such that the table can be understood without having to consult the text. Furthermore, it would be nice if the measurement setup could be briefly described, e.g., how many measurements were taken for each data point, if outliers were removed etc.

j) p. 6, sec. 5.5, tab. 2: For easier visual comparison of the performance numbers, it would be helpful to right-align all time measurements and print them with the same number of digits after the decimal point.

k) p. 6, sec. 6: "... allow to generate regularizer**s** which have ..."

l) p. 7f., sec. 8: Many of the references are lacking DOIs (as requested by the checklist above). Further, most of them have titles that are not properly capitalized, e.g., "fft" instead of "FFT".
