+++ paper/paper.tex
@@ -65,7 +65,7 @@ The mathematical definition of a convolution is
     (h * S)(\mathbf r) = \int_{-\infty}^{\infty} h(\mathbf r - \mathbf x) \cdot S(\mathbf x) \, \mathrm d \mathbf x.
     \label{eq:conv}
 \end{equation}
-To obtain efficient deconvolution algorithms we also need a efficient way to calculate the convolution operations.
+To obtain efficient deconvolution algorithms we also need an efficient way to calculate the convolution operations.
 Today, convolution operations of spatially small kernels $h$ can be calculated on GPUs very efficiently in real space.
 However, often the blurring kernel $h$ is not spatially small and might be also multi-dimensional. In that case, FFT based convolutions
 outperform sliding kernels. In microscopy and astronomy that is often the case, since a defocus broadens the PSF.
@@ -313,7 +313,7 @@ julia> reg(randn((10,20,30,40)))
 
 \section{Experiments}
     In this section we show some experiments and compare parts of the software to other solutions.
-    All experiments were executed on Ubuntu 20.04 using a AMD Ryzen 5 5600X 6-Core processor (12 threads) with \SI{32}{GB} DDR4 RAM and Julia 1.6.1.
+All experiments were executed on Ubuntu 20.04 using a AMD Ryzen 5 5600X 6-Core processor (12 threads) with \SI{32}{GB} DDR4 RAM and Julia 1.6.2.
     If possible, all threads were used.
     As GPU accelerator we used a GeForce RTX 2060 SUPER with \SI{8}{GB} memory and CUDA 11.3 and CUDA.jl 3.3.1.
     Our experiments additionally made use of the following packages in Julia 1.6.2: ArrayInterface.jl\cite{DifferentialEquations.jl-2017}, 
@@ -340,7 +340,9 @@ julia> reg(randn((10,20,30,40)))
             \caption{1D data which has been convolved with a PSF and was degraded with additive Gaussian noise with $\sigma=1.0$. Note that for the deconvolution we have chosen a position interval which is well within the full series so that there are not wrap-around artifacts in the measurement.}
             \label{fig:1D_dataset}
         \end{figure}
-        On the right hand we can see the results of the following deconvolution:
+        On the right hand we can see the results of the following deconvolution.
+    The first statement applies a total padding of $1$ meaning that the size is doubled and therefore the measured data is padded.
+    As we can see, the blue curve in \autoref{fig:1D_datasetb} does not show these intense spikes at the boundaries as the green curve shows.
         \begin{lstlisting}[language=Julia]
 r_p, o = deconvolution(measured, psf, loss=Gauss(),
      iterations=12, regularizer=TV(num_dims=1),
@@ -349,8 +351,6 @@ r, o = deconvolution(measured, psf, loss=Gauss(),
      iterations=12, regularizer=TV(num_dims=1),
      λ=0.05, padding=0)
         \end{lstlisting}
-    The first statement applies a total padding of $1$ meaning that the size is doubled and therefore the measured data is padded.
-    As we can see, the blue curve in \autoref{fig:1D_datasetb} does not show these intense spikes at the boundaries as the green curve shows.
     Also we notice that the deconvolved data in \autoref{fig:1D_datasetb} does not show all the details which are visible in the ground truth.
     The reason is that the PSF is a low pass filter and therefore only frequencies which have been passed through the filter 
     can be enhanced in terms of contrast in a deconvolved result.
@@ -359,43 +359,52 @@ r, o = deconvolution(measured, psf, loss=Gauss(),
         In this part we want to demonstrate the influence of the regularizer.
[…]
