+++ paper/ref.bib
@@ -6,7 +6,8 @@
   number={1},
   pages={65--98},
   year={2017},
-  publisher={SIAM}
+  publisher={SIAM},
+  doi={10.1137/141000671}
 }
 
 @article{Richardson:72,
@@ -43,16 +44,18 @@ abstract = {An iterative method of restoring degraded images was developed by tr
   title={Extrapolation, Interpolation, and Smoothing of Stationary Time Series, with Engineering Applications},
   author={Wiener, Nobert},
   year={1949},
-  publisher={Wiley}
+  publisher={Wiley},
+  doi={2946.001.0001}
 }
 
 
 @InProceedings{Kruse_2017_ICCV,
 author = {Kruse, Jakob and Rother, Carsten and Schmidt, Uwe},
-title = {Learning to Push the Limits of Efficient FFT-Based Image Deconvolution},
+title = {Learning to Push the Limits of Efficient {FFT}-Based Image Deconvolution},
 booktitle = {Proceedings of the IEEE International Conference on Computer Vision (ICCV)},
 month = {Oct},
-year = {2017}
+year = {2017},
+doi={10.1109/ICCV.2017.491}
 }
 
 
@@ -64,7 +67,8 @@ year = {2017}
   volume =	 93,
   number =	 2,
   pages =	 {216--231},
-  note =	 {Special issue on ``Program Generation, Optimization, and Platform Adaptation''}
+  note =	 {Special issue on ``Pr1ogram Generation, Optimization, and Platform Adaptation''},
+  doi = {10.1109/JPROC.2004.840301}
 }
 
 @article{mogensen2018optim,
@@ -93,7 +97,7 @@ year = {2017}
 
 @article{Zygote.jl-2018,
   author    = {Michael Innes},
-  title     = {Don't Unroll Adjoint: Differentiating SSA-Form Programs},
+  title     = {Don't Unroll Adjoint: Differentiating {SSA}-Form Programs},
   journal   = {CoRR},
   volume    = {abs/1810.07951},
   year      = {2018},
@@ -102,7 +106,8 @@ year = {2017}
   eprint    = {1810.07951},
   timestamp = {Tue, 30 Oct 2018 20:39:56 +0100},
   biburl    = {https://dblp.org/rec/bib/journals/corr/abs-1810-07951},
-  bibsource = {dblp computer science bibliography, https://dblp.org}
+  bibsource = {dblp computer science bibliography, https://dblp.org},
+  doi = {10.48550/arXiv.1810.07951}
 }
 
 
@@ -121,7 +126,8 @@ journal = {Math. Program.},
 month = aug,
 pages = {503–528},
 numpages = {26},
-keywords = {conjugate gradient method, Large scale nonlinear optimization, partitioned quasi-Newton method, limited memory methods}
+keywords = {conjugate gradient method, Large scale nonlinear optimization, partitioned quasi-Newton method, limited memory methods},
+doi = {10.1007/BF01589116}
 }
 
 
@@ -153,7 +159,8 @@ abstract = {We present efficient algorithms for image restoration by means of Go
  publisher = {[Oxford University Press, Biometrika Trust]},
  title = {Nonparametric Roughness Penalties for Probability Densities},
  volume = {58},
- year = {1971}
+ year = {1971},
+ doi = {10.2307/2334515},
 }
 
 
@@ -177,14 +184,16 @@ abstract = {We present efficient algorithms for image restoration by means of Go
     volume="115",
     number="",
     pages="28--41",
+    doi = {10.1016/j.ymeth.2016.12.015},
     month="February 15,",
     note=""
 }
 
 
 @article{ikoma2018convex,
-  title={A convex 3D deconvolution algorithm for low photon count fluorescence imaging},
[…]
