### 2022-08-11 #99 @roflmaostc
Hi,
thanks for the detailed feedback! I incorporated it and hopefully it's more clear what we were trying to say

f) this seems to be an issue with @whedon in combination with pgfpots again. Locally, the axis is ok.
j) We tried to always show three significant digits but I see the point of the alignment. I'm not sure how to properly solve that.

l) the automatic capitalization seems to happen by the bibtex settings. I tried to fix with with `{FFT}` in the `ref.bib`
