### 2022-08-11 #99 @roflmaostc
This is how it looks on my Ubuntu machine.
[paper.pdf](https://github.com/JuliaCon/proceedings-review/files/9309306/paper.pdf)
