### 2022-08-11 #99 @sloede
> f) this seems to be an issue with @whedon in combination with pgfpots again. Locally, the axis is ok.

We have had a similar issue with our submission last year. Maybe you can just export the files as PDFs and include them? Also, the code listing is still distributed over 2 pages.
