### 2022-08-11 #99 @roflmaostc
Fixed the code listing.
But what puzzles me, is that I already exported that figure as [pdf](https://github.com/roflmaostc/DeconvOptim.jl/blob/master/paper/figures/1d_plot_data_blurred.pdf) and don't use pgfplots for those plots :confused:
