### 2022-08-11 58dcaec
Convert figures to PDF to avoid including issue with whedon [skip ci]
