### 2022-08-11 #99 @roflmaostc
The bot also compiles the Bibliography slight differently and hence there is suddenly a wraparound
