### 2022-08-11 #99 @sloede
> f) this seems to be an issue with @whedon in combination with pgfpots again. Locally, the axis is ok.

It looks ok now.

> j) We tried to always show three significant digits but I see the point of the alignment. I'm not sure how to properly solve that.

Not sure why you'd require three significant digits for the measurement and wouldn't allow more. Shouldn't a measurement with a given temporal accuracy yield more significant digits if the measured time is increased? That is, if I measure with a sub-millisecond resolution, shouldn't I be able to show measurements with a fixed amount of digits after the decimal point (e.g., 2), no matter if the total interval is one second or 1000 seconds? Anyhow, it's ok as it is if you want to leave it this way.

> l) the automatic capitalization seems to happen by the bibtex settings. I tried to fix with with `{FFT}` in the `ref.bib`

References 2 and 3 are duplicated. Ref 7 should read "Program Generation", not "Pr1ogram Generation". Ref 17 still has capitalization errors. None of the references is equipped with a DOI, as required by the checklist [above](https://github.com/JuliaCon/proceedings-review/issues/99#issue-1290470648).
