### 2022-08-12 5fc0f4c
Remove duplicate reference [skip ci]
