+++ paper/ref.bib
@@ -67,7 +67,7 @@ doi={10.1109/ICCV.2017.491}
   volume =	 93,
   number =	 2,
   pages =	 {216--231},
-  note =	 {Special issue on ``Pr1ogram Generation, Optimization, and Platform Adaptation''},
+  note =	 {Special issue on ``Program Generation, Optimization, and Platform Adaptation''},
   doi = {10.1109/JPROC.2004.840301}
 }
 
@@ -278,15 +278,6 @@ abstract = {The ability of iteratively constrained maximum likelihood (ML) decon
  journal       = {Advances in Engineering Software},
  title         = {Rapid software prototyping for heterogeneous and distributed platforms},
  doi = {10.1016/j.advengsoft.2019.02.002}
-}
-
-@article{LLVM.jl-2017,
- doi           = {10.1109/TPDS.2018.2872064},
- author        = {Besard, Tim and Foket, Christophe and De Sutter, Bjorn},
- issn          = {1045-9219},
- year          = {2018},
- journal       = {IEEE Transactions on Parallel and Distributed Systems},
- title         = {Effective Extensible Programming: Unleashing {Julia} on {GPUs}}
 }
 
 
