### 2022-08-12 #99 @roflmaostc
> References 2 and 3 are duplicated. Ref 7 should read "Program Generation", not "Pr1ogram Generation". Ref 17 still has capitalization errors. None of the references is equipped with a DOI, as required by the checklist [above](https://github.com/JuliaCon/proceedings-review/issues/99#issue-1290470648).

Well spotted, that was generated by [PkgCite.jl](https://github.com/SebastianM-C/PkgCite.jl) but I didn't check it properly, thanks!

I'm not sure why the DOI is not printed but I included it for all references: https://github.com/roflmaostc/DeconvOptim.jl/blob/master/paper/ref.bib

I couldn't spot special settings in recent papers but maybe my `juliacon.cls` is old?
