### 2022-08-12 #99 @roflmaostc
Ah, your [juliacon.bst](https://github.com/trixi-framework/paper-2021-juliacon/blob/main/paper/juliacon.bst) looks different.
