### 2022-09-28 21147c2
Include SnoopPrecompile to speed up TTFP
