+++ paper/ref.bib
@@ -311,3 +311,12 @@ abstract = {The ability of iteratively constrained maximum likelihood (ML) decon
  volume        = {5}
 }
 
+@article{LLVM.jl-2017,
+  author    = {Besard, Tim and Foket, Christophe and De Sutter, Bjorn},
+  title     = {Effective Extensible Programming: Unleashing {Julia} on {GPUs}},
+  journal   = {IEEE Transactions on Parallel and Distributed Systems},
+  year      = {2018},
+  doi       = {10.1109/TPDS.2018.2872064},
+  ISSN      = {1045-9219},
+}
+
