### 2022-11-14 #99 @roflmaostc
Thanks for your valuable input!

> Sorry about the delay, here is my review: Notation: detail each term in the beginning (what is r, I).

done

>
> The mathematical definition of a convolution: convolution here is defined on a univariate domain, it would be good to: define the domain, what S and h are here.

I'm not quite sure what to put here since this seems to be rather lengthy?
https://en.wikipedia.org/wiki/Convolution#Domain_of_definition

>
> "If one applies a FFT on a finite data set", not sure one would consider one image as a data set, on a finite object maybe?

changed it
>
> The experiment summarized in Table 1 seems a bit limited as a comparison, it would be better to have several comparisons aggregated for different sizes (it could be that some methods perform better on small arrays compared to larger ones). This is especially critical since one of the contributions claimed in the abstract is about performance of DeconvOptim.

I could add another size? Maybe a `300x300` 2D array?

>
> The LLVM citation seems broken "LLMV [?]"

fixed.

>
> In the paragraph starting with "As typical for Julia packages", it would be good to replace "code" with more specific terms. "all code is written in" -> the entire package is written in Julia.

>
> Overall, I'm not sure the sentence on typical Julia packages being written in Julia is necessary here. It also implies that FFTW and CUDA are the only packages that do not contain only pure Julia.

>
> Does the package allow for other optimizing methods from Optim.jl? Have the authors tried some of them?

Yes we, have a keyword `opt` to specify the optimizer. For convolution, the optimizer `LBFGS` seems to be fastest and preferred option in the community. We can only confirm this. [See here](https://en.wikipedia.org/wiki/Limited-memory_BFGS#Applications)

>
> Instead of detailing the packages used for the experiments, list what they were used for.

Honestly, I generated the citation with [PkgCite.jl](https://github.com/SebastianM-C/PkgCite.jl) and some of the packages are just dependencies from other packages.
Sometimes, I can't tell exactly what's their functionality in the full code base.

>
> The lines of Figure 6b are too thin to read well.

Agree, happened because I had to change everything to .png instead of pgfplots.
Fixed

>
> Fig 8 takes up a lot of space, could it be resized to fit images two by two?
>

I agree, but I couldn't find a better solution given the image width.

> Double-check the grammar, there were minor improvements and corrections to do: "spatially based convolutions" -> spatial-based?
>
> "In many frameworks such" -> "In many frameworks, such"
>
> "Apart from forward model" -> "Apart from the forward model"
>
> "A (smoothed) Total variation" -> A (smoothed) Total Variation
>
Fixed them

> The authors advocate for letting autodiff compute gradients for flexibility. Have they measured the impact of autodiff versus manual (in-place) gradient computations? Could the package let the user opt out of AD by providing a gradient themselves?

Currently not directly possible, but there is the more flexible method `generic_invert` where one could use own forward models and register the gradients with `ChainRules`. That works and I did that in the past.
