### 2022-11-16 #99 @roflmaostc
I don't really know what's going on with this bot but the compiled layout is different on my machine....
[paper.pdf](https://github.com/JuliaCon/proceedings-review/files/10025805/paper.pdf)
