### 2022-11-17 #99 @matbesancon
> Honestly, I generated the citation with [PkgCite.jl](https://github.com/SebastianM-C/PkgCite.jl) and some of the packages are just dependencies from other packages.
Sometimes, I can't tell exactly what's their functionality in the full code base.

I can understand the utility to extract them like that but I think it really should be written and phrased properly afterwards or remove them if there is nothing directly relevant about them for the paper and/or package. People can always look up the package and see the dependencies for themselves
