### 2022-11-21 fa43c56
Update references section
