### 2022-11-21 362a741
Fix style of BenchmarkTools.jl [skip ci]
