### 2022-11-21 #99 @roflmaostc
Yeah, I agree to that.

I removed the sentence since most of the important packages were already cited before.

@vchuravy I still don't really know what's going on with @whedon. Which LaTeX version is it using? Is it a really old one? The compiled PDFs looks considerably different to the ones on my machine (`pdfTeX 3.14159265-2.6-1.40.20 (TeX Live 2019/Debian)`)
