### 2022-11-21 #99 @roflmaostc
Interesting, this one looks very similar to the one on my machine :D
