### 2023-01-03 62ef3fd
first attempt to streamline the regularizers
