+++ Project.toml
@@ -21,7 +21,7 @@ Zygote = "e88e6eb3-aa80-5325-afca-941959d7151f"
 
 [compat]
 ChainRulesCore = "1"
-FFTW = "1.2, 1.3, 1.4"
+FFTW = "1.2, 1.3, 1.4, 1.5"
 FillArrays = "0.12, 0.13"
 Interpolations = "0.12.10, 0.13, 0.14"
 LineSearches = "7.1"
@@ -32,7 +32,7 @@ SnoopPrecompile = "1"
 StatsBase = "0.33"
 Tullio = "0.3"
 Zygote = "0.6"
-julia = "1, 1.5, 1.6, 1.7, 1.8"
+julia = "1, 1.6, 1.7, 1.8, 1.9"
 
 [extras]
 ImageMagick = "6218d12a-5da1-5696-b52f-db25d2ecc6d1"
