+++ v0.7.0/search_index.js
@@ -1,0 +1,3 @@
+var documenterSearchIndex = {"docs":
+[{"location":"references/#References","page":"References","title":"References","text":"","category":"section"},{"location":"references/","page":"References","title":"References","text":"See here for a list of references","category":"page"},{"location":"references/","page":"References","title":"References","text":"","category":"page"},{"location":"function_references/loss/#Loss-Functions","page":"Loss Functions","title":"Loss Functions","text":"","category":"section"},{"location":"function_references/loss/","page":"Loss Functions","title":"Loss Functions","text":"Poisson\npoisson_aux\nGauss\ngauss_aux\nScaledGauss\nscaled_gauss_aux","category":"page"},{"location":"function_references/loss/#DeconvOptim.Poisson","page":"Loss Functions","title":"DeconvOptim.Poisson","text":"Poisson()\n\nReturns a function to calculate Poisson loss Check the help of poisson_aux.\n\n\n\n\n\n","category":"function"},{"location":"function_references/loss/#DeconvOptim.poisson_aux","page":"Loss Functions","title":"DeconvOptim.poisson_aux","text":"poisson_aux(μ, meas, storage=similar(μ))\n\nCalculates the Poisson loss for μ and meas. μ can be of larger size than meas. In that case we extract a centered region from μ of the same size as meas.\n\n\n\n\n\n","category":"function"},{"location":"function_references/loss/#DeconvOptim.Gauss","page":"Loss Functions","title":"DeconvOptim.Gauss","text":"Gauss()\n\nReturns a function to calculate Gauss loss. Check the help of gauss_aux.\n\n\n\n\n\n","category":"function"},{"location":"function_references/loss/#DeconvOptim.gauss_aux","page":"Loss Functions","title":"DeconvOptim.gauss_aux","text":"gauss_aux(μ, meas, storage=similar(μ))\n\nCalculates the Gauss loss for μ and meas. μ can be of larger size than meas. In that case we extract a centered region from μ of the same size as meas.\n\n\n\n\n\n","category":"function"},{"location":"function_references/loss/#DeconvOptim.ScaledGauss","page":"Loss Functions","title":"DeconvOptim.ScaledGauss","text":"ScaledGauss()\n\nReturns a function to calculate scaled Gauss loss. Check the help of scaled_gauss_aux.\n\n\n\n\n\n","category":"function"},{"location":"function_references/loss/#DeconvOptim.scaled_gauss_aux","page":"Loss Functions","title":"DeconvOptim.scaled_gauss_aux","text":"scaled_gauss_aux(μ, meas, storage=similar(μ); read_var=0)\n\nCalculates the scaled Gauss loss for μ and meas. read_var=0 is the readout noise variance of the sensor. μ can be of larger size than meas. In that case we extract a centered region from μ of the same size as meas.\n\n\n\n\n\n","category":"function"},{"location":"function_references/regularizer/#Regularizers","page":"Regularizers","title":"Regularizers","text":"","category":"section"},{"location":"function_references/regularizer/#CPU","page":"Regularizers","title":"CPU","text":"","category":"section"},{"location":"function_references/regularizer/","page":"Regularizers","title":"Regularizers","text":"TV\nTikhonov\nGR\nTH\nHS","category":"page"},{"location":"function_references/regularizer/#DeconvOptim.TV","page":"Regularizers","title":"DeconvOptim.TV","text":"TV(; <keyword arguments>)\n\nThis function returns a function to calculate the Total Variation regularizer of a n-dimensional array. \n\nArguments\n\nnum_dims=2: \nsum_dims=1:num_dims: A array containing the dimensions we want to sum over\nweights=nothing: A array containing weights to weight the contribution of    different dimensions. If weights=nothing all dimensions are weighted equally.\nstep=1: A integer indicating the step width for the 
[…]
