### 2023-06-20 #99 @matbesancon
@roflmaostc you will need to create a zenodo archive of the repository, once this is done can you post the DOI here?
