### 2023-07-18 ed07c60
fix: add notes to documentation about shift
[Ticket: #44]
