### 2023-07-18 4e4a9e3
fixup! fix: add notes to documentation about shift
