### 2023-07-18 707a6e5
fix: add notes to documentation about shift
[Ticket: #44]
