### 2023-09-03 #99 @matbesancon
@roflmaostc I believe the part to blame for the cryptic message above is the fact that one DOI seems invalid
