### 2023-09-03 #99 @roflmaostc
I don't think that JAX has a DOI:
https://github.com/google/jax/blob/main/CITATION.bib
