+++ paper/ref.bib
@@ -45,7 +45,7 @@ abstract = {An iterative method of restoring degraded images was developed by tr
   author={Wiener, Nobert},
   year={1949},
   publisher={Wiley},
-  doi={2946.001.0001}
+  doi={10.7551/mitpress/2946.003.0001}
 }
 
 
