### 2023-09-03 #99 @matbesancon
@roflmaostc the paper is here: https://github.com/JuliaCon/proceedings-papers/blob/jcon.00099/jcon.00099/10.21105.jcon.00099.pdf

one thing maybe, could we have the code block on a single page instead of cut in two?
