### 2023-09-03 4f80a93
Move figure [skip ci]
